Added CAPTURED_ADDRESS to handle offline usage
In an offline hotspot, we want people to be prompted with the portal
when going to external addresses.
For this, we'd serve DNS for all addresses they request.
If we'd return our own IP, we'd end-up going to our home webservice. Not what we want.
We'd thus return them another IP, that's not on the LAN, so they have to go through
our server (router).
There, we intercept those packets (because we know that CAPTURED_ADDRESS) and redirect
them to the captive portal.
People being already registered, they'd be greated with the already-registered page
that tells them to go to the proper home URL.